Crate swap_buffer_queue

swap-buffer-queue

A buffering MPSC queue.

This library is intended to be a (better, I hope) alternative to traditional MPSC queues in the context of a buffering consumer, by moving the buffering part directly into the queue.

It is especially well suited for IO writing workflow, see write and write_vectored.

The crate is no_std (some buffer implementations may require std).

Examples

// Initialize the queue with a capacity
let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
// Enqueue some values
queue.try_enqueue(0).unwrap();
queue.try_enqueue(1).unwrap();
// Dequeue a slice to the enqueued values
let slice = queue.try_dequeue().unwrap();
assert_eq!(slice.deref(), &[0, 1]);
// Enqueued values can also be retrieved
assert_eq!(slice.into_iter().collect::<Vec<_>>(), vec![0, 1]);

Modules

asyncasync

Asynchronous implementation of SBQueue.

buffer

Buffer definition and simple implementations.

error

Queue error types.

notify

Tool for (a)synchronous implementation.

syncsync

Synchronous implementation of SBQueue.

writewrite

Buffer implementations to be used with Write::write.

write_vectoredwrite-vectored

Buffer implementations to be used with Write::write_vectored.

Structs

SBQueue

A buffered MPSC “swap-buffer” queue.

Type Definitions

AsyncSBQueue

An asynchronous implementation of SBQueue.

SyncSBQueue

A synchronous implementation of SBQueue.