Trait swap_buffer_queue::buffer::Buffer

pub unsafe trait Buffer<T>: Default {
    type Slice<'a>
    where
        Self: 'a;

    fn value_size(value: &T) -> usize;
    fn capacity(&self) -> usize;
    fn debug(&self, debug_struct: &mut DebugStruct<'_, '_>);
    unsafe fn insert(&mut self, index: usize, value: T);
    unsafe fn slice(&mut self, len: usize) -> Self::Slice<'_>;
    unsafe fn clear(&mut self, len: usize);
}

SBQueue buffer.

Safety

This trait is meant to be used inside an UnsafeCell, that’s why mutability of self parameter doesn’t follow safe Rust rules.

capacity and debug methods must be safe to call concurrently to other methods receiving &mut self, e.g. insert.

insert method may be called multiple times concurrently

The rest behaves normally, i.e. slice method borrow a mutable reference, which must be released to call clear, etc.

Required Associated Types

type Slice<'a>
where
Self: 'a

The slice type returned by slice method.

Required Methods

fn value_size(value: &T) -> usize

Return the size taken by a value in the buffer.

fn capacity(&self) -> usize

Returns the buffer’s capacity.

fn debug(&self, debug_struct: &mut DebugStruct<'_, '_>)

Formats the buffer fields in debugging context.

unsafe fn insert(&mut self, index: usize, value: T)

Inserts value into the buffer at the given index.

Safety

Every call to this method must have a non overlapping half-open interval [index,index+Buffer::value_size(&value)).

unsafe fn slice(&mut self, len: usize) -> Self::Slice<'_>

Returns a slice of the buffer.

Safety

Half-open interval [0,len) must have been inserted (see insert) before calling this method.

unsafe fn clear(&mut self, len: usize)

Clears the buffer.

Safety

Half-open interval [0,len) must have been inserted (see insert) before calling this method.

Calling this method clears the inserted value, i.e. inserted interval is reset to [0, 0) (see insert), meaning new value can be inserted.