Trait swap_buffer_queue::buffer::Drainable

pub unsafe trait Drainable<T>: Buffer<T> {
    type Drain<'a>: Iterator<Item = T>
    where
        Self: 'a;

    unsafe fn drain(&mut self, len: usize) -> Self::Drain<'_>;
}

Buffer whose value can be removed and returned as an iterator.

Safety

The trait extends Buffer contract, with drain method following the same rule as Buffer::clear. It means that Buffer::capacity can be called concurrently, but not Buffer::insert or Buffer::clear.

Required Associated Types

type Drain<'a>: Iterator<Item = T>
where
Self: 'a

The drain type returned by drain.

It may implement `Iterator<Item=T>’.

Required Methods

unsafe fn drain(&mut self, len: usize) -> Self::Drain<'_>

Removes all elements of the buffer and returns them as an iterator.

Safety

Half-open interval [0,len) must have been inserted (see Buffer::insert) before calling this method.

The iterator returned must be exhausted; it will have the same effect than calling Buffer::clear