Struct swap_buffer_queue::SBQueue

pub struct SBQueue<B, T, N = ()>where
    B: Buffer<T>,{ /* private fields */ }

A buffered MPSC “swap-buffer” queue.

Implementations

impl<B, T, N> SBQueue<B, T, N>where
B: Buffer<T>,
N: Default,

pub fn new() -> Self

Create a new queue using buffer default.

Buffer default may have a non-zero capacity, e.g. array buffer.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::new();

impl<B, T, N> SBQueue<B, T, N>where
B: Buffer<T> + Resizable<T>,
N: Default,

pub fn with_capacity(capacity: usize) -> Self

Creates a new queue with the given capacity.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);

impl<B, T, N> SBQueue<B, T, N>where
B: Buffer<T>,

pub fn notify(&self) -> &N

Create a new queue with the given capacity.

Examples

use swap_buffer_queue::notify::Notify;

let queue: AsyncSBQueue<VecBuffer<usize>, usize> = AsyncSBQueue::with_capacity(42);
queue.notify().notify_dequeue();

pub fn capacity(&self) -> usize

Returns the current buffer capacity.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
assert_eq!(queue.capacity(), 42);

pub fn len(&self) -> usize

Returns the current buffer length.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
assert_eq!(queue.len(), 0);
queue.try_enqueue(0).unwrap();
assert_eq!(queue.len(), 1);

pub fn is_empty(&self) -> bool

Returns true if the current buffer is empty.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
assert!(queue.is_empty());

pub fn is_closed(&self) -> bool

Returns true if the queue is closed.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
assert!(!queue.is_closed());
queue.close();
assert!(queue.is_closed());

pub fn reopen(&self)

Reopen a closed queue.

Calling this method when the queue is not closed has no effect.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
queue.close();
assert!(queue.is_closed());
queue.reopen();
assert!(!queue.is_closed());

impl<B, T, N> SBQueue<B, T, N>where
B: Buffer<T>,
N: Notify,

pub fn try_enqueue(&self, value: T) -> Result<(), TryEnqueueError<T>>

Tries enqueuing the given value into the queue.

Enqueuing will fail if the queue has insufficient capacity, or if it is closed. In case of success, it will notify waiting dequeuing operations using Notify::notify_dequeue.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(1);
queue.try_enqueue(0).unwrap();
// queue is full
assert_eq!(
    queue.try_enqueue(0),
    Err(TryEnqueueError::InsufficientCapacity(0))
);
// let's close the queue
queue.close();
assert_eq!(queue.try_enqueue(0), Err(TryEnqueueError::Closed(0)));

pub fn try_dequeue(&self) -> Result<BufferSlice<'_, B, T, N>, TryDequeueError>

Tries dequeuing a buffer with all enqueued values from the queue.

This method swaps the current buffer with the other one, which is empty. All concurrent insertions must end before the the current buffer is really dequeuable, so the queue may be in a transitory state where try_dequeue must be retried. In this state, after a spin loop, this method will return a TryDequeueError::Pending error.

Dequeuing also fails if the queue is empty, or if it is closed. Moreover, as the algorithm is MPSC, dequeuing is protected against concurrent calls, failing with TryDequeueError::Conflict error.

It returns a BufferSlice, which holds, as its name may indicate, a reference to the dequeued buffer. That’s why, the concurrent dequeuing protection is maintained for the lifetime of the buffer slice.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
queue.try_enqueue(0).unwrap();
queue.try_enqueue(1).unwrap();
let slice = queue.try_dequeue().unwrap();
assert_eq!(slice.deref(), &[0, 1]);
// dequeuing cannot be done concurrently (`slice` is still in scope)
assert_eq!(queue.try_dequeue().unwrap_err(), TryDequeueError::Conflict);
drop(slice);
// let's close the queue
queue.try_enqueue(2).unwrap();
queue.close();
// queue can be dequeued while closed when not empty
let slice = queue.try_dequeue().unwrap();
assert_eq!(slice.deref(), &[2]);
drop(slice);
assert_eq!(queue.try_dequeue().unwrap_err(), TryDequeueError::Closed)

pub fn close(&self)

Closes the queue.

Closed queue can no more accept enqueuing, but it can be dequeued while not empty. Calling this method on a closed queue has no effect. See reopen to reopen a closed queue.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(42);
queue.try_enqueue(0).unwrap();
queue.close();
assert!(queue.is_closed());
assert_eq!(queue.try_enqueue(1), Err(TryEnqueueError::Closed(1)));
assert_eq!(queue.try_dequeue().unwrap().deref(), &[0]);
assert_eq!(queue.try_dequeue().unwrap_err(), TryDequeueError::Closed);

impl<B, T, N> SBQueue<B, T, N>where
B: Buffer<T> + Resizable<T>,
N: Notify,

pub fn try_dequeue_and_resize(
&self,
capacity: usize,
insert: Option<impl Iterator<Item = T>>
) -> Result<BufferSlice<'_, B, T, N>, TryDequeueError>

Tries dequeuing a buffer with all enqueued values from the queue, and resizes the next buffer to be used for enqueuing.

This method is an extension of try_dequeue method. In fact, before swapping the buffers, next one is empty and protected, so it can be resized, and it is also possible to add values in it before making it available for enqueuing. This can be used to make the queue unbounded.

It is worth to be noted that buffer is resized even if the queue is empty; in this case, buffer will be swapped, whereas it would not be the case with try_dequeue method call on an empty queue.

Examples

let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::with_capacity(1);
queue.try_enqueue(0).unwrap();
// queue is full
assert_eq!(
    queue.try_enqueue(1),
    Err(TryEnqueueError::InsufficientCapacity(1))
);
// dequeue and resize, inserting elements before the buffer is available
let slice = queue
    .try_dequeue_and_resize(3, Some(std::iter::once(42)))
    .unwrap();
assert_eq!(slice.deref(), &[0]);
drop(slice);
// capacity has been increased
queue.try_enqueue(1).unwrap();
queue.try_enqueue(2).unwrap();
let slice = queue.try_dequeue().unwrap();
assert_eq!(slice.deref(), &[42, 1, 2]);

An amortized unbounded recipe

fn enqueue_unbounded<T>(
    queue: &SBQueue<VecBuffer<T>, T>,
    overflow: &Mutex<Vec<T>>,
    mut value: T,
) -> Result<(), EnqueueError<T>> {
    // first, try to enqueue normally
    match queue.try_enqueue(value) {
        Ok(()) => return Ok(()),
        Err(TryEnqueueError::Closed(v)) => return Err(EnqueueError(v)),
        Err(TryEnqueueError::InsufficientCapacity(v)) => value = v,
    };
    // if the enqueuing fails, lock the overflow
    let mut guard = overflow.lock().unwrap();
    // retry to enqueue (we never know what happened during lock acquisition)
    match queue.try_enqueue(value) {
        Ok(()) => return Ok(()),
        Err(TryEnqueueError::Closed(v)) => return Err(EnqueueError(v)),
        Err(TryEnqueueError::InsufficientCapacity(v)) => value = v,
    };
    // then push the values to the overflow vector
    guard.push(value);
    // notify possible waiting dequeue
    queue.notify().notify_dequeue();
    Ok(())
}

fn try_dequeue_unbounded<'a, T>(
    queue: &'a SBQueue<VecBuffer<T>, T>,
    overflow: &Mutex<Vec<T>>,
) -> Result<BufferSlice<'a, VecBuffer<T>, T, ()>, TryDequeueError> {
    // lock the overflow and use `try_dequeue_and_resize` to drain the overflow into the
    // queue
    let mut guard = overflow.lock().unwrap();
    queue.try_dequeue_and_resize(queue.capacity() + guard.len(), Some(guard.drain(..)))
}

// queue is initialized with zero capacity
let queue: SBQueue<VecBuffer<usize>, usize> = SBQueue::new();
let overflow = Mutex::new(Vec::new());
assert_eq!(queue.capacity(), 0);
enqueue_unbounded(&queue, &overflow, 0).unwrap();
assert_eq!(queue.capacity(), 0);
assert_eq!(
    try_dequeue_unbounded(&queue, &overflow).unwrap_err(),
    TryDequeueError::Empty
);
assert_eq!(queue.capacity(), 1);
assert_eq!(queue.len(), 1);
enqueue_unbounded(&queue, &overflow, 1).unwrap();
assert_eq!(queue.capacity(), 1);
assert_eq!(queue.len(), 1);
assert_eq!(overflow.lock().unwrap().len(), 1);
assert_eq!(
    try_dequeue_unbounded(&queue, &overflow).unwrap().deref(),
    &[0]
);
assert_eq!(queue.capacity(), 2);
assert_eq!(queue.len(), 1);
enqueue_unbounded(&queue, &overflow, 2).unwrap();
assert_eq!(
    try_dequeue_unbounded(&queue, &overflow).unwrap().deref(),
    &[1, 2]
);

impl<B, T> SBQueue<B, T, AsyncNotifier>where
B: Buffer<T>,

pub async fn enqueue(&self, value: T) -> Result<(), EnqueueError<T>>

Available on crate feature async only.

Enqueues the given value inside the queue.

This method extends try_enqueue by waiting asynchronously AsyncNotifier::notify_enqueue call, i.e. when a buffer is dequeued, in case of insufficient capacity.

Examples

let queue: Arc<AsyncSBQueue<VecBuffer<usize>, usize>> =
    Arc::new(AsyncSBQueue::with_capacity(1));
queue.try_enqueue(0).unwrap();
assert_eq!(
    queue.try_enqueue(0),
    Err(TryEnqueueError::InsufficientCapacity(0))
);
// queue is full, let's spawn an enqueuing task and dequeue
let queue_clone = queue.clone();
let task = tokio::spawn(async move { queue_clone.enqueue(1).await });
assert_eq!(queue.try_dequeue().unwrap().deref(), &[0]);
// enqueuing task has succeeded
task.await.unwrap().unwrap();
assert_eq!(queue.try_dequeue().unwrap().deref(), &[1]);
// let's close the queue
queue.try_enqueue(2).unwrap();
let queue_clone = queue.clone();
let task = tokio::spawn(async move { queue_clone.enqueue(3).await });
queue.close();
assert_eq!(task.await.unwrap(), Err(EnqueueError(3)));

pub async fn dequeue(
&self
) -> Result<BufferSlice<'_, B, T, AsyncNotifier>, DequeueError>

Available on crate feature async only.

Dequeues a buffer with all enqueued values from the queue.

This method extends try_dequeue by waiting asynchronously AsyncNotifier::notify_dequeue call, i.e. when a value is enqueued, in case of empty queue.

Examples

let queue: Arc<AsyncSBQueue<VecBuffer<usize>, usize>> =
    Arc::new(AsyncSBQueue::with_capacity(1));
assert_eq!(queue.try_dequeue().unwrap_err(), TryDequeueError::Empty);
// queue is empty, let's spawn a dequeuing task and enqueue
let queue_clone = queue.clone();
let task = tokio::spawn(async move {
    Ok::<_, DequeueError>(queue_clone.dequeue().await?.into_iter().collect::<Vec<_>>())
});
queue.try_enqueue(0).unwrap();
// dequeuing task has succeeded
assert_eq!(task.await.unwrap().unwrap().deref(), &[0]);
// let's close the queue
let queue_clone = queue.clone();
let task = tokio::spawn(async move {
    Ok::<_, DequeueError>(queue_clone.dequeue().await?.into_iter().collect::<Vec<_>>())
});
queue.close();
assert_eq!(task.await.unwrap().unwrap_err(), DequeueError::Closed);

impl<B, T> SBQueue<B, T, SyncNotifier>where
B: Buffer<T>,

pub fn try_enqueue_timeout(
&self,
value: T,
timeout: Duration
) -> Result<(), TryEnqueueError<T>>

Available on crate feature sync only.

Tries enqueuing the given value inside the queue with a timeout.

This method extends try_enqueue by waiting synchronously, with a timeout, SyncNotifier::notify_enqueue call, i.e. when a buffer is dequeued, in case of insufficient capacity.

Examples

let queue: Arc<SyncSBQueue<VecBuffer<usize>, usize>> = Arc::new(SyncSBQueue::with_capacity(1));
queue.try_enqueue(0).unwrap();
assert_eq!(
    queue.try_enqueue_timeout(1, Duration::from_millis(1)),
    Err(TryEnqueueError::InsufficientCapacity(1))
);
let queue_clone = queue.clone();
let task = std::thread::spawn(move || {
    std::thread::sleep(Duration::from_millis(1));
    queue_clone.try_dequeue().unwrap();
});
queue
    .try_enqueue_timeout(1, Duration::from_secs(1))
    .unwrap();

pub fn enqueue(&self, value: T) -> Result<(), EnqueueError<T>>

Available on crate feature sync only.

Enqueues the given value inside the queue.

This method extends try_enqueue by waiting synchronously SyncNotifier::notify_enqueue call, i.e. when a buffer is dequeued, in case of insufficient capacity.

Examples

let queue: Arc<SyncSBQueue<VecBuffer<usize>, usize>> = Arc::new(SyncSBQueue::with_capacity(1));
queue.try_enqueue(0).unwrap();
assert_eq!(
    queue.try_enqueue(1),
    Err(TryEnqueueError::InsufficientCapacity(1))
);
// queue is full, let's spawn an enqueuing task and dequeue
let queue_clone = queue.clone();
let task = std::thread::spawn(move || queue_clone.enqueue(1));
assert_eq!(queue.try_dequeue().unwrap().deref(), &[0]);
// enqueuing task has succeeded
task.join().unwrap().unwrap();
assert_eq!(queue.try_dequeue().unwrap().deref(), &[1]);
// let's close the queue
queue.try_enqueue(2).unwrap();
let queue_clone = queue.clone();
let task = std::thread::spawn(move || queue_clone.enqueue(3));
queue.close();
assert_eq!(task.join().unwrap(), Err(EnqueueError(3)));

pub fn try_dequeue_timeout(
&self,
timeout: Duration
) -> Result<BufferSlice<'_, B, T, SyncNotifier>, TryDequeueError>

Available on crate feature sync only.

Tries dequeuing a buffer with all enqueued values from the queue with a timeout.

This method extends try_dequeue by waiting synchronously, with a timeout, SyncNotifier::notify_dequeue call, i.e. when a value is enqueued, in case of empty queue.

Examples

let queue: Arc<SyncSBQueue<VecBuffer<usize>, usize>> = Arc::new(SyncSBQueue::with_capacity(1));
assert_eq!(
    queue
        .try_dequeue_timeout(Duration::from_millis(1))
        .unwrap_err(),
    TryDequeueError::Empty
);
let queue_clone = queue.clone();
let task = std::thread::spawn(move || {
    std::thread::sleep(Duration::from_millis(1));
    queue_clone.try_enqueue(0).unwrap();
});
assert_eq!(
    queue
        .try_dequeue_timeout(Duration::from_secs(1))
        .unwrap()
        .deref(),
    &[0]
);

pub fn dequeue(
&self
) -> Result<BufferSlice<'_, B, T, SyncNotifier>, DequeueError>

Available on crate feature sync only.

Dequeues a buffer with all enqueued values from the queue.

This method extends try_dequeue by waiting synchronously SyncNotifier::notify_dequeue call, i.e. when a value is enqueued, in case of empty queue.

Examples

let queue: Arc<SyncSBQueue<VecBuffer<usize>, usize>> = Arc::new(SyncSBQueue::with_capacity(1));
assert_eq!(queue.try_dequeue().unwrap_err(), TryDequeueError::Empty);
// queue is empty, let's spawn a dequeuing task and enqueue
let queue_clone = queue.clone();
let task = std::thread::spawn(move || {
    Ok::<_, DequeueError>(queue_clone.dequeue()?.into_iter().collect::<Vec<_>>())
});
queue.try_enqueue(0).unwrap();
// dequeuing task has succeeded
assert_eq!(task.join().unwrap().unwrap().deref(), &[0]);
// let's close the queue
let queue_clone = queue.clone();
let task = std::thread::spawn(move || {
    Ok::<_, DequeueError>(queue_clone.dequeue()?.into_iter().collect::<Vec<_>>())
});
queue.close();
assert_eq!(task.join().unwrap().unwrap_err(), DequeueError::Closed);