Module swap_buffer_queue::write_vectored

Available on crate feature write-vectored only.

Buffer implementations to be used with Write::write_vectored.

WriteVectoredArrayBuffer and WriteVectoredVecBuffer allows buffering a slice of IoSlice, saving the cost of dequeuing io-slices one by one to collect them after. (Internally, two buffers are used: one of the values, and one for the io-slices)

Examples

// Creates a WriteVectoredVecBuffer queue
let queue: SBQueue<WriteVectoredVecBuffer<Vec<u8>>, Vec<u8>> = SBQueue::with_capacity(100);
queue.try_enqueue(vec![0; 256]).unwrap();
queue.try_enqueue(vec![42; 42]).unwrap();
let mut slice = queue.try_dequeue().unwrap();
// Adds a header with the total size of the slices
let total_size = (slice.total_size() as u16).to_be_bytes();
let mut frame = slice.frame(.., Some(IoSlice::new(&total_size)), None);
// Let's pretend we have a writer
let mut writer: Vec<u8> = Default::default();
assert_eq!(writer.write_vectored(&mut frame).unwrap(), 300);

Structs

VectoredFrame

A framed part of a VectoredSlice, with an IoSlice header and an IoSlice trailer (see VectoredSlice::frame).

VectoredSlice

A vectored slice, i.e. a slice of IoSlice.

WriteVectoredArrayBuffer

A buffer of IoSlice of size N

WriteVectoredVecBuffer

A buffer of IoSlice